Added in navigation from an edit event view to the correct tab in the events list
This update makes it so that after an event is updated from the event
admin view, it returns the user to the correct tab.

For example, if I'm click on an event to edit from the past view, when
I'm done editing it I will now be navigated back to the past view.